requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.6
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
dataclasses-json==0.6.7
distro==1.9.0
faiss-cpu==1.13.2
fastapi==0.128.1
filelock==3.20.3
frozenlist==1.8.0
fsspec==2026.1.0
gitdb==4.0.12
GitPython==3.1.46
greenlet==3.3.1
groq==0.37.1
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.4.0
idna==3.11
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
langchain==1.2.8
langchain-classic==1.0.1
langchain-community==0.4.1
langchain-core==1.2.8
langchain-groq==1.1.2
langchain-openai==1.1.7
langchain-text-splitters==1.1.0
langgraph==1.0.7
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.7
langgraph-sdk==0.3.3
langsmith==0.6.8
MarkupSafe==3.0.3
marshmallow==3.26.2
mpmath==1.3.0
multidict==6.7.1
mypy_extensions==1.1.0
narwhals==2.16.0
networkx==3.6.1
numpy==2.4.2
openai==2.16.0
orjson==3.11.7
ormsgpack==1.12.2
packaging==26.0
pandas==2.3.3
pillow==12.1.0
propcache==0.4.1
protobuf==6.33.5
pyarrow==23.0.0
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pydeck==0.9.1
pypdf==6.6.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
regex==2026.1.15
requests==2.32.5
requests-toolbelt==1.0.0
rpds-py==0.30.0
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.0
sentence-transformers==5.2.2
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.46
starlette==0.50.0
streamlit==1.54.0
sympy==1.14.0
tenacity==9.1.3
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.2
toml==0.10.2
torch==2.10.0
tornado==6.5.4
tqdm==4.67.3
transformers==5.0.0
typer-slim==0.21.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uuid_utils==0.14.0
uvicorn==0.40.0
watchdog==6.0.0
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
